backend/requirements.txt
fastapi
uvicorn
rank-bm25
tqdm
transformers
openai
pymysql
python-dotenv